Changed type of input numbers to 'long long'
* use 'AC_PROG_CC_C99' in configure.ac

+++ include/rgxg/number.h
@@ -28,13 +28,13 @@
 /* needed for rgxg_options_t */
 #include "types.h"
 
-int rgxg_number(long number, int base, char* regex,
+int rgxg_number(long long number, int base, char* regex,
         rgxg_options_t options);
 
-int rgxg_number_range(long first, long last, int base, char* regex,
+int rgxg_number_range(long long first, long long last, int base, char* regex,
         rgxg_options_t options);
 
-int rgxg_number_greaterequal(long number, int base, char* regex,
+int rgxg_number_greaterequal(long long number, int base, char* regex,
         rgxg_options_t options);
 
 #endif /* _RGXG_NUMBER_H_INCLUDED */
